Abstract row navigation into a hook
